Reserve a static buffer in memory and expose it via VESC_IF
Moves data around in RAM to free up a continuous region of 62kB. To free
up space:
- Lisp heap is reduced in size from ~19kB to ~9kB.
- C heap is reduced to ~2.4kB. It's basically only used in blackmagic
  (used for flashing stuff), where it allocates at most 1kB.
- ADC sample buffers are reduced from 1600 samples to 8. These are used
  in VESC Tool desktop for sampling and plotting.

This region is exposed via the last 12 bytes of the memory dedicated for
the VESC_IF package interface, the layout is as follows:
magic [4B]
data* [4B]
length [4B]

The magic number is set to a specific value as a safety check. If
interpreted as an address, it would point into unused memory on
STM32F40xxx. If a package expecting this data in the last 12 bytes of
VESC_IF would be loaded on a firmware that contains real interface, it
should fail this check and keep working normally.

The data pointer points to the start of the available memory range
of the given length.