IMU: Rework reading out values for LSM6DS3
The code was reading out IMU register values at the configured rate
without any synchronization and without checking if there are new values
in the registers, thus it was subject to (at least):
- Aliasing
- Skpping a sample
- Reading a sample twice

This should really be solved via an interrupt from the IMU. In absence
of that, the code in this commit checks the presence of new data in the
IMU registers and attempts to plan the polling so that it's a bit early
(rather than late) by subtracting 2 ticks from the next expected sample
time. If there is no new sample when checked, it only sleeps for 1 tick
to intercept the new sample as soon as possible.

For completeness, there is also a big issue playing a role here, the I2C
communication being done on the CPU. That means excessive CPU load and
the reading of IMU values being slow, and the communication being subject
to interruption by interrupts or other threads. This has a big impact on
the process of retrieving IMU samples and nothing can be done about that
besides using a HW peripheral for the communication.