IMU: Rework reading out values for LSM6DS3
The code reading out IMU values was problematic in two aspects:

- Lack of synchronization: It's attempting to read samples at the same
  frequency that is configured for the IMU ODR, but this is subject to
  timing issues / jitter and missing or double-reading samples,
  depending on the actual frequencies.

- Running the communication on the CPU: This causes high CPU load and
  doesn't allow to reach higher sampling rates as the CPU saturates very
  early.

Due to the implementation of the sleep delay, for higher frequencies
the real frequency ends up being lower than configured and it's
effectively capped. E.g. for Sample Rate 833Hz the real frequency is
around 550-650 Hz.

Now the IMU sampling always runs at the highest frequency (6660Hz) and
the Sample Rate is now used only to determine the frequency for polling
the IMU.

Filtering options are limited for this frequency, the filtering is set
as follows (for the -TRC variant):
- LOW: Accelerometer 400Hz, Gyro 351Hz
- MEDIUM: Accelerometer 133.2Hz, Gyro 237Hz
- HIGH: Accelerometer 66.6Hz, Gyro 173Hz

For the non-TRC version the gyro doesn't support ODR higher than 1660 Hz
and the filtering is tied to the ODR, so on non-TRC the old mechanism of
setting ODR according to the Sample Rate is preserved.

Feature: Rework the sampling of the LSM6DS3 IMU
 >
 Fixes the timing when reading out the IMU samples and together with
 increasing the communication frequency allows significantly higher IMU
 sample rates.
 >
 The filtering configuration has also been changed and is now as follows:
 >
 - LOW: Accelerometer 400Hz, Gyro 351Hz
 >
 - MEDIUM: Accelerometer 133.2Hz, Gyro 237Hz
 >
 - HIGH: Accelerometer 66.6Hz, Gyro 173Hz
 >
 Recommended settings:
 >
 With these changes it's recommended to set IMU Sample Rate to 1200 Hz.
 >
 For filtering it's recommended to run the highest possible filtering
 frequency which is lower than half your IMU Sample Rate by a margin
 (say, 2.5-3 times lower).
 >
 That is, if you run at Sample Rate of 1200 Hz, set LOW filtering.
 >
 Note on the non-TRC version of LSM6DS3 (irrelevant to most Onewheel
 controllers):
 >
 Majority of recent/relevant VESC controllers (all Onewheel controllers
 to the extent of my knowledge) use the -TRC version of the LSM6DS3 IMU.
 For the non-TRC version, the possibilities are limited (the filter
 frequencies are different than listed above) and as of writing this
 text the code was not tested on a non-TRC IMU (I don't have the
 hardware). It's probably not worth supporting the non-TRC version
 anymore, but an effort was made to keep it functional.